docs/12/MUIT Chapter7/genai-agent-advanced-book/chapter7/pyproject.toml
[project]
name = "chapter7"
version = "0.1.0"
description = "現場で活用するためのAIエージェント実践入門 第7章"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>1.68",
    "langgraph==0.2.14",
    "pydantic-settings==2.4.0",
    "langchain>=0.2.15,<0.3.0",
    "langchain-openai>=0.1.25",
    "python-dotenv>=1.1.1",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.11.4",
    "ipykernel==6.29.5",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
select = [
    "F", # Flake8
    "B", # Black
    "I", # isort
    "E", # error
    "W"  # warning
]
ignore = []
fixable = ["ALL"]
exclude = [
    ".ruff_cache"
]
line-length = 119
indent-width = 4
target-version = "py312"

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401", "F403"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.isort]
force-single-line = false
